requirements.txt
python-dotenv
langgraph
langchain 
langchainhub
langchain_google_genai
langchain_community